Fix repeated loading of CL-CUDA without an SDK.
Previously, when there was no cuda sdk found and a successfully loaded
CL-CUDA adsf system was loaded again, then the grovel file was not
compiled, but the files that followed it in the system definition
were.

With this change, the grovel file is not included in the sytem
definition if the sdk is not found, so this problem doesn't arise.